More refactoring on integration package, adding suite test for db and helpers
