+++ db/database_suite_test.go
@@ -1,0 +1,13 @@
+package db
+
+import (
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+
+	"testing"
+)
+
+func TestEncoders(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "DB Suite")
+}
