+++ encoders/encoders_suite_test.go
@@ -1,0 +1,13 @@
+package encoders
+
+import (
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+
+	"testing"
+)
+
+func TestEncoders(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Encoders Suite")
+}
