+++ integration/config.json
@@ -1,0 +1,4 @@
+{
+  "SWAP_DIRECTORY": "/tmp/",
+  "LOGFILE": "/dev/null"
+}
